Refactor question storage
