+++ teacher_interface.py
@@ -1,19 +1,4 @@
-import json
-import os
-
-QUESTIONS_FILE = 'questions.json'
-
-
-def load_questions():
-    if not os.path.exists(QUESTIONS_FILE):
-        return []
-    with open(QUESTIONS_FILE, 'r', encoding='utf-8') as f:
-        return json.load(f)
-
-
-def save_questions(qs):
-    with open(QUESTIONS_FILE, 'w', encoding='utf-8') as f:
-        json.dump(qs, f, indent=2)
+from question_store import load_questions, save_questions
 
 
 def add_question():
